requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
attrs==23.2.0
black==24.4.2
certifi==2024.6.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
eval_type_backport==0.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy==1.10.0
mypy-extensions==1.0.0
orjson==3.10.5
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pycodestyle==2.11.1
pydantic==2.7.4
pydantic_core==2.18.4
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.2.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
starlette==0.37.2
trio==0.25.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0